+++ app/controllers/refinery/admin/blog/comments_controller.rb
@@ -2,37 +2,44 @@ module Refinery
   module Admin
     module Blog
       class CommentsController < ::Admin::BaseController
+        
+        cache_sweeper Refinery::BlogSweeper
 
         crudify :'refinery/blog_comment',
                 :title_attribute => :name,
                 :order => 'published_at DESC'
 
         def index
-          @blog_comments = Refinery::BlogComment.unmoderated
+          @blog_comments = Refinery::BlogComment.unmoderated.page(params[:page])
+          
           render :action => 'index'
         end
 
         def approved
           unless params[:id].present?
-            @blog_comments = Refinery::BlogComment.approved
+            @blog_comments = Refinery::BlogComment.approved.page(params[:page])
+            
             render :action => 'index'
           else
             @blog_comment = Refinery::BlogComment.find(params[:id])
             @blog_comment.approve!
-            flash[:notice] = t('approved', :scope => 'admin.blog.comments', :author => @blog_comment.name)
-            redirect_to :action => params[:return_to] || 'index'
+            flash[:notice] = t('approved', :scope => 'refinery.admin.blog.comments', :author => @blog_comment.name)
+            
+            redirect_to main_app.url_for(:action => params[:return_to] || 'index')
           end
         end
 
         def rejected
           unless params[:id].present?
-            @blog_comments = Refinery::BlogComment.rejected
+            @blog_comments = Refinery::BlogComment.rejected.page(params[:page])
+            
             render :action => 'index'
           else
             @blog_comment = Refinery::BlogComment.find(params[:id])
             @blog_comment.reject!
-            flash[:notice] = t('rejected', :scope => 'admin.blog.comments', :author => @blog_comment.name)
-            redirect_to :action => params[:return_to] || 'index'
+            flash[:notice] = t('rejected', :scope => 'refinery.admin.blog.comments', :author => @blog_comment.name)
+            
+            redirect_to main_app.url_for(:action => params[:return_to] || 'index')
           end
         end
