+++ app/controllers/refinery/blog/posts_controller.rb
@@ -1,6 +1,9 @@
 module Refinery
   module Blog
     class PostsController < BlogController
+      
+      caches_page :index
+      # cache_sweeper Refinery::BlogSweeper, :only => [:comment]
 
       before_filter :find_all_blog_posts, :except => [:archive]
       before_filter :find_blog_post, :only => [:show, :comment, :update_nav]
@@ -38,10 +41,10 @@ module Refinery
 
           if Refinery::BlogComment::Moderation.enabled?
             flash[:notice] = t('thank_you_moderated', :scope => 'blog.posts.comments')
-            redirect_to blog_post_url(params[:id])
+            redirect_to main_app.blog_post_url(params[:id])
           else
             flash[:notice] = t('thank_you', :scope => 'blog.posts.comments')
-            redirect_to blog_post_url(params[:id],
+            redirect_to main_app.blog_post_url(params[:id],
                                       :anchor => "comment-#{@blog_comment.to_param}")
           end
         else
