+++ app/views/refinery/admin/blog/comments/_sortable_list.html.erb
@@ -1,7 +1,7 @@
 <ul id='sortable_list'>
   <%= render :partial => 'comment', :collection => @blog_comments %>
 </ul>
-<%= render :partial => "/shared/admin/sortable_list",
+<%= render :partial => "/refinery/admin/sortable_list",
            :locals => {
              :continue_reordering => (defined?(continue_reordering) ? continue_reordering : true)
            } %>
