+++ app/views/refinery/admin/blog/settings/notification_recipients.html.erb
@@ -14,7 +14,7 @@
     <%= t('.example') %>
   </p>
 
-  <%= render :partial => "/shared/admin/form_actions",
+  <%= render :partial => "/refinery/admin/form_actions",
              :locals => {
                :f => nil,
                :continue_editing => false,
